Move signup Google fragment to login library
